+++ packages/pluggableWidgets/combobox-web/src/Combobox.editorConfig.ts
@@ -23,7 +23,9 @@ export function getProperties(values: ComboboxPreviewProps, defaultProperties: P
             "optionsSourceAssociationCustomContentType",
             "optionsSourceAssociationDataSource",
             "selectedItemsStyle",
-            "selectionMethod"
+            "selectionMethod",
+            "selectAllButton",
+            "selectAllButtonCaption"
         ]);
         if (values.optionsSourceType === "boolean") {
             hidePropertiesIn(defaultProperties, values, ["clearable"]);
@@ -51,6 +53,10 @@ export function getProperties(values: ComboboxPreviewProps, defaultProperties: P
 
         if (values.showFooter === false) {
             hidePropertiesIn(defaultProperties, values, ["menuFooterContent"]);
+        }
+
+        if (values.selectAllButton === false) {
+            hidePropertiesIn(defaultProperties, values, ["selectAllButtonCaption"]);
         }
     }